Structure refactor * Removed unnecessary files from boilerplate * Removed karma libraries and configuration, no usage * Moved all configurations to infrastructure layer to try keeping root directory clean
