--- index.js
@@ -1,3 +1,0 @@
-// install babel hooks in the main process
-require('babel-register');
-require('./main.js');
