+++ src/openMarket/infrastructure/configuration/server.js
@@ -10,7 +10,7 @@ import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
 import { spawn } from 'child_process';
 
-import config from './webpack.config.development';
+import config from './webpack/webpack.config.development';
 
 const argv = require('minimist')(process.argv.slice(2));
