+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.base.js
@@ -4,7 +4,7 @@
 
 import path from 'path';
 import validate from 'webpack-validator';
-import { dependencies as externals } from './app/package.json';
+import { dependencies as externals } from '../../../../../app/package.json';
 
 export default validate({
   module: {
