+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.electron.js
@@ -16,7 +16,7 @@ export default validate(merge(baseConfig, {
   // 'main.js' in root
   output: {
     path: __dirname,
-    filename: './app/main.js'
+    filename: '../../../../../app/main.js'
   },
 
   plugins: [
