+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.eslint.js
@@ -1,0 +1,3 @@
+require('babel-register');
+
+module.exports = require('./webpack.config.development.js');
