+++ src/openMarket/infrastructure/configuration/webpack/webpack.config.production.js
@@ -17,8 +17,8 @@ export default validate(merge(baseConfig, {
   entry: ['babel-polyfill', './app/index'],
 
   output: {
-    path: path.join(__dirname, 'app/dist'),
-    publicPath: '../dist/'
+    path: path.join(__dirname, '../../../../../app/dist'),
+    publicPath: '../../../../../dist/'
   },
 
   module: {
