--- view/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Hello WORLD!</title>
-</head>
-<body>
-    <p>Blaaaa blaaa</p>
-</body>
-</html>
