--- webpack.config.eslint.js
@@ -1,3 +1,0 @@
-require('babel-register');
-
-module.exports = require('./webpack.config.development');
